feat: configuring new html report generator, start working with publishArtifact mechanism

+++ sysdig-scan-task/src/index.ts
@@ -5,6 +5,8 @@ import fs = require('fs');
 import { InputFetch } from './InputFetch';
 import { getScanningEngine, buildScanningEngineArg, runScanningEngine, runScanningEnginev2 } from './ScanningEngine';
 import { dockerComandArg, scanningImage, buildInlineScanCommand, runInlineScanner } from './InlineScanner';
+import { SysdigScan, generateHTMLTableFromSysdigJSON } from './ReportGenerator';
+// import { publishArtifact } from './PublishArtifact';
 
 async function run() {
   try {
@@ -19,13 +21,19 @@ async function run() {
 
       // runScanningEngine(scanningEngine);
       await runScanningEnginev2(scanningEngine);
-
-      console.log('[INFO] Generating HTML report');
-      var jsonData = JSON.parse(fs.readFileSync(fetch.jsonOutputFile, 'utf8'));
-      var htmlTable = generateHTMLTableFromJSON(jsonData);
-      fs.writeFileSync('output.html', htmlTable);
-
-      console.log('[INFO] HTML report generated successfully');
+      if (fs.existsSync(fetch.jsonOutputFile)) {
+        console.log('[INFO] Generating HTML report');
+        var jsonData: SysdigScan = JSON.parse(fs.readFileSync(fetch.jsonOutputFile, 'utf8'));
+        var htmlTable = generateHTMLTableFromSysdigJSON(jsonData);
+        fs.writeFileSync('output.html', htmlTable);
+        console.log('[INFO] HTML report generated successfully');
+      } else {
+        console.log('File not found:', fetch.jsonOutputFile);
+      }
+      if (fetch.autoPublishArtifacts == true) {
+        console.log('[INFO] Auto publish artifacts not available in this version. Please follow instructions provided on README.md to configure your own publishing mechanism.');
+        // publishArtifact('html_report', 'output.html');
+      }
     }
     else {
       console.log('[INFO] Starting Sysdig Inline Scanner');
@@ -43,17 +51,6 @@ async function run() {
     }
     tl.setResult(tl.TaskResult.Failed, errorMessage);
   }
-}
-
-// We treat jsonData as any so we can have flexibility in the JSON structure
-function generateHTMLTableFromJSON(jsonData: any) {
-  var packages = jsonData.result.packages.filter(function (pkg: any) { return pkg.suggestedFix; });
-  var htmlContent = '<table><tr><th>Package Name</th><th>Package Type</th><th>Version</th><th>Suggested Fix</th></tr>';
-  packages.forEach(function (pkg: any) {
-      htmlContent += "<tr>\n            <td>".concat(pkg.name, "</td>\n            <td>").concat(pkg.type, "</td>\n            <td>").concat(pkg.version, "</td>\n            <td>").concat(pkg.suggestedFix, "</td>\n        </tr>");
-  });
-  htmlContent += '</table>';
-  return htmlContent;
 }
 
 run();
